requirements.txt
Backtesting==0.3.3
bokeh==3.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
idna==3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pandas-ta==0.3.14b0
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tornado==6.4
tzdata==2024.1
urllib3==2.2.1
xyzservices==2024.4.0
beautifulsoup4==4.12.3
soupsieve==2.5
algo_trading_helpers @ file:///Users/dhavalchheda/Desktop/Study/python/packages/algo_trading_helpers/dist/algo_trading_helpers-0.2-py3-none-any.whl#sha256=b1dc99bcb6ba66530b61c7cc5cf01b8b8b8f480d987e70de43616d1cde8daa27

algo_trading_helpers @ file:///Users/dhavalchheda/Desktop/Study/python/packages/algo_trading_helpers/dist/algo_trading_helpers-0.2-py3-none-any.whl#sha256=b1dc99bcb6ba66530b61c7cc5cf01b8b8b8f480d987e70de43616d1cde8daa27
Backtesting==0.3.3
beautifulsoup4==4.12.3
bokeh==3.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
idna==3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pandas-ta==0.3.14b0
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
schedule==1.2.1
six==1.16.0
soupsieve==2.5
tornado==6.4
tzdata==2024.1
urllib3==2.2.1
xyzservices==2024.4.0
algo_trading_helpers @ file:///Users/dhavalchheda/Desktop/Study/python/packages/algo_trading_helpers/dist/algo_trading_helpers-0.2-py3-none-any.whl#sha256=0d7801d255a7b86656248a82fe78b48faf998acbe257ed13429f7d0c4b4f788e
Backtesting==0.3.3
beautifulsoup4==4.12.3
bokeh==3.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
idna==3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pandas-ta==0.3.14b0
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
schedule==1.2.1
six==1.16.0
soupsieve==2.5
tornado==6.4
tzdata==2024.1
urllib3==2.2.1
xyzservices==2024.4.0
